Remove unnecessary code generation of the Contract class for interface types
